\name{penaltySVM-internal}
\alias{.core}
\alias{.calc.mult.inv_Q_mat2}
\alias{.correctness}
\alias{.create.covariance.matrix}
\alias{.EstNuLong}
\alias{.EstNuShort}
\alias{.extend.to.quad.matrix}
\alias{.find.inverse}
%\alias{.Random.seed}
\alias{.required}
\alias{.run.cv}
\alias{.whatiscorrect}

\title{Internal penaltySVM objects}
\description{Internal penaltySVM objects.}
\details{These are not to be called by the user.}
\keyword{internal}