\name{penalizedSVM-package}
\alias{penalizedSVM-package}
\alias{penalizedSVM}
\docType{package}
\title{
Feature Selection SVM using Penalty Functions  
}
\description{
Feature Selection SVM using penalty functions. The smoothly clipped absolute deviation (SCAD) and L1-norm penalties are 
availible up to now. Other functions will be implemented in the near feature.  
}

\details{
\tabular{ll}{
Package: \tab penaltySVM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-12-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

The main function is \code{\link{svm.fs}, see the documentation file with examples}
}
\author{ Natalia Becker, Axel Benner, Wiebke Werft 

Maintainer: Natalia Becker (natalia.becker@dkfz.de)
}

\references{
Zhang, H. H., Ahn, J., Lin, X. and Park, C. (2006). \emph{Gene selection using
support vector machines with nonconvex penalty.} Bioinformatics, \bold{22}, pp. 88-95. 

Fung, G. and Mangasarian, O. L. (2004). \emph{ A feature selection newton method for 
support vector machine classification.} Computational Optimization and Applications Journal ,\bold{28.2} , pp. 185-202.

}


\keyword{ package }
